Fix broken snooze forever button on org dashboard
/cc @getsentry/ui
